+++ src/pages/Goals.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+export default class Goals extends Component { 
+  state = { 
+  }
+    render () {                                   
+       return (
+         <div id='container'>
+           
+         </div>
+       )
+    }
+ }
